requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
boto3==1.37.33
botocore==1.37.33
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.2
Django==5.2
django-cleanup==9.0.0
django-filter==25.1
django-storages==1.14.6
django-stubs-ext==5.1.3
djangorestframework==3.16.0
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.3
Markdown==3.7
mypy==1.15.0
mypy-extensions==1.0.0
pem==23.1.0
pika==1.3.2
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250402
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
